Thu 7 Dec 5:25 PM

+++ database/migrations/2023_09_14_121017_create_suppliers_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('phone');
             $table->foreignId('governorate_id')->constrained('governorates')->onDelete('cascade');
             $table->string('address');
-            $table->decimal('balance', 20 )->default(0);
+            $table->bigInteger('balance')->default(0);
             $table->text('note')->nullable();
             $table->softDeletes();
             $table->timestamps();
