+++ database/migrations/2023_09_17_125839_create_products_table.php
@@ -15,11 +15,10 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('code')->unique()->nullable();
-            // the barcode should be not nullable
             $table->string('barcode')->unique()->nullable();
             $table->integer('quantity')->default(0);
-            $table->decimal('price', 10 )->default(0);
-            $table->decimal('cost', 10 )->default(0);
+            $table->bigInteger('price')->default(0);
+            $table->bigInteger('cost')->default(0);
             $table->text('note')->nullable();
             $table->softDeletes();
             $table->timestamps();
