+++ database/migrations/2023_09_18_123656_create_purchase_raw_material_table.php
@@ -16,12 +16,12 @@ return new class extends Migration
             $table->foreignId('purchase_id')->constrained()->onDelete('cascade');
             $table->foreignId('raw_material_id')->constrained()->onDelete('cascade');
             $table->integer('quantity');
-            $table->decimal('subtotal', 10 );
-            $table->decimal('total', 10 );
-            $table->decimal('unit_price', 10 );
-            $table->decimal('discount_amount', 10 )-> nullable();
-            $table->decimal('discount_percentage', 5 )-> nullable();
-            $table->decimal('cost', 10 )->default(0);
+            $table->bigInteger('subtotal');
+            $table->bigInteger('total');
+            $table->bigInteger('unit_price');
+            $table->bigInteger('discount_amount')-> nullable();
+            $table->bigInteger('discount_percentage')-> nullable();
+            $table->bigInteger('cost')->default(0);
             $table->softDeletes();
             $table->timestamps();
         });
