+++ database/migrations/2023_09_27_120551_create_expenses_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
             $table->string('title');
             $table->string('description')->nullable();
-            $table->decimal('amount', 10);
+            $table->bigInteger('amount');
             $table->softDeletes();
             $table->timestamps();
         });
